chore: cleanup repo (#13)
* local-ic cleanup

* add POA test

* cleanup misc go and proto

* cleanup

* undo accidental params change (!=0)

* remove handled todos

* lint

* proto-gen latest

* release docker image on tag
