+++ .golangci.yml
@@ -40,7 +40,9 @@ linters-settings:
       - prefix(github.com/cosmos)
       - prefix(cosmossdk.io)
       - prefix(github.com/cosmos/cosmos-sdk)
-      - prefix(github.com/CosmosContracts/juno)
+      - prefix(github.com/strangelove-ventures/poa)
+      - prefix(github.com/reecepbcups/tokenfactory)
+      - prefix(github.com/liftedinit/manifest-ledger)
   gosec:
     excludes:
       - G404
