+++ app/upgrades/noop/upgrades.go
@@ -3,12 +3,12 @@ package noop
 import (
 	"context"
 
-	"github.com/liftedinit/manifest-ledger/app/upgrades"
-
 	storetypes "cosmossdk.io/store/types"
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 
 	"github.com/cosmos/cosmos-sdk/types/module"
+
+	"github.com/liftedinit/manifest-ledger/app/upgrades"
 )
 
 // NewUpgrade constructor
