+++ proto/manifest/v1/tx.proto
@@ -19,7 +19,7 @@ service Msg {
   // Since: cosmos-sdk 0.47
   rpc UpdateParams(MsgUpdateParams) returns (MsgUpdateParamsResponse);
 
-  // manual payout mint
+  // PayoutStakeholders allows the authority to manually pay out stakeholders.
   rpc PayoutStakeholders(MsgPayoutStakeholders) returns (MsgPayoutStakeholdersResponse);
 }
 
@@ -44,7 +44,7 @@ message MsgUpdateParams {
 // Since: cosmos-sdk 0.47
 message MsgUpdateParamsResponse {}
 
-// MsgPayoutStakeholders manually pays out previously saved stakeholders.
+// MsgPayoutStakeholders is the Msg/PayoutStakeholders request type.
 message MsgPayoutStakeholders {
   option (cosmos.msg.v1.signer) = "authority";
   option (gogoproto.equal) = false;
@@ -52,6 +52,7 @@ message MsgPayoutStakeholders {
   // authority is the address of the controlling account.
   string authority = 1 [(cosmos_proto.scalar) = "cosmos.AddressString"];
 
+  // payout is the amount of tokens paid to the current stakeholders.
   cosmos.base.v1beta1.Coin payout = 2 [
     (gogoproto.nullable)     = false,
     (amino.dont_omitempty)   = true,
@@ -60,5 +61,5 @@ message MsgPayoutStakeholders {
   ];
 }
 
-// MsgPayoutStakeholdersResponse
+// MsgPayoutStakeholdersResponse defines the response structure for executing a MsgPayoutStakeholders message.
 message MsgPayoutStakeholdersResponse {}
