+++ x/manifest/README.md
@@ -1,3 +1,3 @@
 # x/manifest
 
-Just a simple extension to toggle if the TokenFactory can mint tokens or not.
+The manifest module handles automatic & manual minting of a set token to stakeholders.
