+++ x/manifest/keeper/keeper.go
@@ -3,8 +3,6 @@ package keeper
 import (
 	"context"
 
-	"github.com/liftedinit/manifest-ledger/x/manifest/types"
-
 	"cosmossdk.io/collections"
 	storetypes "cosmossdk.io/core/store"
 	"cosmossdk.io/log"
@@ -14,6 +12,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	mintkeeper "github.com/cosmos/cosmos-sdk/x/mint/keeper"
+
+	"github.com/liftedinit/manifest-ledger/x/manifest/types"
 )
 
 type Keeper struct {
@@ -51,10 +51,10 @@ func NewKeeper(
 		mintKeeper: mintKeeper,
 		bankKeeper: bankKeeper,
 
-		authority: authority,
-
 		// Stores
 		Params: collections.NewItem(sb, types.ParamsKey, "params", codec.CollValue[types.Params](cdc)),
+
+		authority: authority,
 	}
 
 	schema, err := sb.Build()
