+++ x/manifest/keeper/keeper_test.go
@@ -3,15 +3,15 @@ package keeper_test
 import (
 	"testing"
 
-	"github.com/liftedinit/manifest-ledger/app"
-	appparams "github.com/liftedinit/manifest-ledger/app/params"
 	"github.com/stretchr/testify/suite"
-
-	"github.com/cometbft/cometbft/crypto/ed25519"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
+
+	"github.com/liftedinit/manifest-ledger/app"
+	"github.com/liftedinit/manifest-ledger/app/apptesting"
+	appparams "github.com/liftedinit/manifest-ledger/app/params"
 )
 
 type testFixture struct {
@@ -36,20 +36,9 @@ func initFixture(t *testing.T) *testFixture {
 		GRPCQueryRouter: s.App.GRPCQueryRouter(),
 		Ctx:             s.Ctx,
 	}
-	s.TestAccs = CreateRandomAccounts(3)
+	s.TestAccs = apptesting.CreateRandomAccounts(3)
 
 	s.EncodingCfg = encCfg
 
 	return &s
-}
-
-// CreateRandomAccounts is a function return a list of randomly generated AccAddresses
-func CreateRandomAccounts(numAccts int) []sdk.AccAddress {
-	testAddrs := make([]sdk.AccAddress, numAccts)
-	for i := 0; i < numAccts; i++ {
-		pk := ed25519.GenPrivKey().PubKey()
-		testAddrs[i] = sdk.AccAddress(pk.Address())
-	}
-
-	return testAddrs
 }
