+++ x/manifest/module.go
@@ -6,10 +6,6 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	manifestv1 "github.com/liftedinit/manifest-ledger/api/manifest/v1"
-	"github.com/liftedinit/manifest-ledger/x/manifest/client/cli"
-	"github.com/liftedinit/manifest-ledger/x/manifest/keeper"
-	"github.com/liftedinit/manifest-ledger/x/manifest/types"
 	"github.com/spf13/cobra"
 
 	abci "github.com/cometbft/cometbft/abci/types"
@@ -25,6 +21,11 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	mintkeeper "github.com/cosmos/cosmos-sdk/x/mint/keeper"
+
+	manifestv1 "github.com/liftedinit/manifest-ledger/api/manifest/v1"
+	"github.com/liftedinit/manifest-ledger/x/manifest/client/cli"
+	"github.com/liftedinit/manifest-ledger/x/manifest/keeper"
+	"github.com/liftedinit/manifest-ledger/x/manifest/types"
 )
 
 const (
