+++ x/manifest/types/errors.go
@@ -4,7 +4,4 @@ import (
 	"cosmossdk.io/errors"
 )
 
-var (
-	ErrGettingMinter         = errors.Register(ModuleName, 1, "getting minter in ante handler")
-	ErrManualMintingDisabled = errors.Register(ModuleName, 2, "manual minting is disabled due to automatic inflation being on")
-)
+var ErrManualMintingDisabled = errors.Register(ModuleName, 1, "manual minting is disabled due to automatic inflation being on")
